refactor: deconstruct god module coordinator.py
- Moved data transformation, model mapping, and registry checks into `UpdateProcessor`.
- Consolidated logic from `MerakiDataProcessor` into `UpdateProcessor` and deleted `data_processor.py`.
- Encapsulated raw API fetching and top-level timeout logic in `DataFetchManager`.
- Delegated failure handling and interval management to `UpdateProcessor`.
- Reduced `coordinator.py` to a thin orchestrator (~170 lines).
- Ensured all tests pass and handle specific error messages correctly.
